// api/generated/model/apiChainRecordJSON.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * ChainCoord API
 * Chain launch coordination server. Coordinators and validators authenticate via secp256k1 challenge-response (ADR-036)
 * OpenAPI spec version: 1.0
 */

export interface ApiChainRecordJSON {
  bech32_prefix?: string;
  binary_name?: string;
  binary_sha256?: string;
  binary_version?: string;
  chain_id?: string;
  chain_name?: string;
  denom?: string;
  genesis_time?: string;
  gentx_deadline?: string;
  max_commission_change_rate?: string;
  max_commission_rate?: string;
  min_self_delegation?: string;
  min_validator_count?: number;
  repo_commit?: string;
  repo_url?: string;
  total_supply?: string;
}